Fix build due to ut64 ≡ uint64_t

+++ src/widgets/BreakpointWidget.cpp
@@ -78,7 +78,7 @@ QVariant BreakpointModel::data(const QModelIndex &index, int role) const
     case Qt::EditRole:
         switch (index.column()) {
         case AddrColumn:
-            return breakpoint.addr;
+            return (quint64)breakpoint.addr;
         case TraceColumn:
             return breakpoint.trace;
         case EnabledColumn:
